implement swagger

+++ users/views/instructor_views.py
@@ -2,10 +2,12 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
 from users.models import Instructor
+from users.permissions import IsInstructorOrAdmin
 from users.serializers import InstructorSerializer, InstructorUpdateSerializer
-
+#===============================all instructors========================
 class InstructorListView(APIView):
-     def get(self, request):
+#  permission_classes = [IsInstructorOrAdmin]
+ def get(self, request):
         instructors = Instructor.objects.select_related("user").all()
         serializer = InstructorSerializer(instructors, many=True)
         return Response(
@@ -16,8 +18,9 @@ class InstructorListView(APIView):
             status=status.HTTP_200_OK
         )
 # ================= profile / update / delete  instructor=================
-class InstructorDetailView(APIView):
 
+class InstructorDetailView(APIView):
+    permission_classes = [IsInstructorOrAdmin]
     def get_object(self, slug):
         return Instructor.objects.select_related("user").filter(user__slug=slug).first()
 
@@ -28,6 +31,7 @@ class InstructorDetailView(APIView):
                 {"message": "Instructor not found"},
                 status=status.HTTP_404_NOT_FOUND
             )
+        self.check_object_permissions(request, instructor)
         serializer = InstructorSerializer(instructor)
         return Response(
             {"message": "Instructor retrieved successfully", "data": serializer.data},
@@ -41,8 +45,10 @@ class InstructorDetailView(APIView):
                 {"message": "Instructor not found"},
                 status=status.HTTP_404_NOT_FOUND
             )
+        self.check_object_permissions(request, instructor)
         serializer = InstructorUpdateSerializer(instructor, data=request.data, partial=True)
         if serializer.is_valid():
+
             serializer.save()
             return Response(
                 {"message": "Instructor updated successfully", "data": serializer.data},
@@ -56,7 +62,8 @@ class InstructorDetailView(APIView):
             return Response(
                 {"message": "Instructor not found"},
                 status=status.HTTP_404_NOT_FOUND
-            )
+            ) 
+        self.check_object_permissions(request, instructor)
         instructor.delete()
         return Response(
             {"message": "Instructor deleted successfully"},
